Attempt 2 of the Google Kick Start 2019 D3 Challenge
Attempted the challenge at https://codingcompetitions.withgoogle.com/kickstart/round/0000000000051061/0000000000161476.

Passed the smaller test case, making me part of the 18% of participants to do so. Timed out on the larger test case.

Placed 1819 out of 4397 total participants overall in Google Kickstart 2019 Round D.

+++ GoogleKickStart2019D3/src/googlekickstart2019d3/Solution.java
@@ -29,8 +29,8 @@ public class Solution {
         return new Scanner(System.in);
     }
     
-    private static int getCost(int[] spots, int[] costs, int numberOfStalls) {
-        int minCost = Integer.MAX_VALUE;
+    private static long getCost(int[] spots, int[] costs, int numberOfStalls) {
+        long minCost = Long.MAX_VALUE;
         for (int j = 0; j < spots.length; j++) {
             int[] modCosts = new int[spots.length - 1];
             for (int i = 0; i < j; i++) {
@@ -40,7 +40,7 @@ public class Solution {
                 modCosts[i - 1] = costs[i] + Math.abs(spots[j] - spots[i]);
             }
             Arrays.sort(modCosts);
-            int currentCost = costs[j];
+            long currentCost = costs[j];
             for (int i = 0; i < numberOfStalls; i++) {
                 currentCost += modCosts[i];
             }
